BoardDisplayer need refactory
